+++ src/codec.ts
@@ -12,16 +12,22 @@ export function isCodec(obj: any): boolean {
         && isFunction(obj.encode);
 }
 
-export function getEncoder<T,U>(obj: Codec<T,U>): Converter<T,U>{
-    if (!isCodec(obj))
-        throw new TypeError(`object is not a codec: ${obj}`);
-    return obj.encode.bind(obj);
+export function getEncoder<T,U>(obj: Codec<T,U> | Converter<T,U>): Converter<T,U>{
+    if (!isCodec(obj)) {
+        if (isFunction(obj))
+            return obj as Converter<T,U>;
+        throw new TypeError(`object is not a codec or converter: ${obj}`);
+    }
+    return (obj as Codec<T,U>).encode.bind(obj) as Converter<T,U>;
 }
 
-export function getDecoder<T,U>(obj: Codec<T,U>): Converter<U,T> {
-    if (!isCodec(obj))
+export function getDecoder<T,U>(obj: Codec<T,U> | Converter<U,T>): Converter<U,T> {
+    if (!isCodec(obj)) {
+        if (isFunction(obj))
+            return obj as Converter<U,T>;
         throw new TypeError(`object is not a codec: ${obj}`);
-    return obj.decode.bind(obj);
+    }
+    return (obj as Codec<T,U>).decode.bind(obj);
 
 }