+++ src/decorators.ts
@@ -1,4 +1,4 @@
-import {isBlank, isDefined} from "./lang";
+import {isBlank, isDefined, Type, Constructor} from "./lang";
 
 export class DecoratorError extends Error {
     constructor(errMessage) {
@@ -36,22 +36,35 @@ class InstanceCache {
  * will be cached on the object under the `__instanceCache` attribute.
  */
 export function memoize() {
-    function decorate(target: Object, propName: string | symbol, descriptor: TypedPropertyDescriptor<any>) {
-        var originalMethod = descriptor.value;
-        if (originalMethod.length > 0)
-            throw new DecoratorError(`@memoize: Cannot apply to a method with arguments`);
-        var instanceCache = new InstanceCache();
-
-        // Important: Do not use fat arrow syntax. We want the context
-        // of `this` in the endpoint to be the same as the instance.
-        descriptor.value = function () {
-            var cache = InstanceCache.getCache(this);
+    function cacheValue(propName: string | symbol, getterOrMethod: () => any): () => any {
+        return function () {
+            let cache = InstanceCache.getCache(this);
             if (!cache.has(propName)) {
-                cache.set(propName, originalMethod.apply(this));
+                cache.set(propName, getterOrMethod.apply(this));
             }
             return cache.get(propName);
-        };
+        }
+    }
+
+    function decorate(target: Object, propName: string | symbol) {
+        let descriptor = Object.getOwnPropertyDescriptor(target, <string>propName);
+        let getterOrMethod: () => any;
+
+        if (!isDefined(descriptor))
+            throw new DecoratorError(`@memoize: Can only be applied to methods or properties`);
+
+        if (isDefined(descriptor.get)) {
+            descriptor.get = cacheValue(propName, descriptor.get);
+        } else {
+            let method = descriptor.value;
+            if (!isDefined(method.length))
+                throw new DecoratorError(`@memoize: \'${propName} is not a method`);
+            if (method.length > 0)
+                throw new DecoratorError(`@memoize: Method \'${propName} should not take arguments`);
+            descriptor.value = cacheValue(propName, method);
+        }
         return descriptor;
     }
     return decorate;
 }
+